MDPS/requirements.txt
# MDPS - Market Data Processing System
# Core dependencies for the complete system

# Core Python packages
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine Learning and Data Science
scikit-learn>=1.0.0
scikit-optimize>=0.9.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
tensorflow>=2.8.0
torch>=1.12.0
transformers>=4.20.0
optuna>=2.10.0

# Deep Learning and Neural Networks
keras>=2.8.0
tensorflow-gpu>=2.8.0  # Optional: for GPU acceleration
pytorch-lightning>=1.6.0

# Data Processing and Analysis
polars>=0.15.0  # Fast DataFrame library
vaex>=4.13.0   # Big data processing
dask>=2022.1.0  # Parallel computing
ray>=1.13.0    # Distributed computing

# Financial and Trading
yfinance>=0.1.70
ccxt>=2.0.0    # Cryptocurrency exchange library
alpha-vantage>=2.3.1
quandl>=3.6.0
ta>=0.10.0     # Technical analysis
talib-binary>=0.4.0  # Technical analysis library
finta>=1.3      # Financial technical analysis

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
bokeh>=2.4.0
altair>=4.2.0
holoviews>=1.15.0
dash>=2.3.0    # Web dashboard framework

# Database and Storage
sqlalchemy>=1.4.0
aiosqlite>=0.17.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
redis>=4.0.0
influxdb>=5.3.0
clickhouse-driver>=0.2.0
mongodb>=4.0.0

# Web and API
fastapi>=0.78.0
uvicorn>=0.17.0
aiohttp>=3.8.0
requests>=2.28.0
websockets>=10.3
flask>=2.2.0
django>=4.0.0

# Async and Concurrency
asyncio-mqtt>=0.11.0
aiofiles>=0.8.0
asyncio-throttle>=1.0.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=0.19.0
configparser>=5.2.0
click>=8.1.0
typer>=0.4.0

# Logging and Monitoring
structlog>=21.5.0
loguru>=0.6.0
prometheus-client>=0.14.0
statsd>=3.3.0

# Testing and Development
pytest>=7.0.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0
pytest-mock>=3.7.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0

# Performance and Profiling
memory-profiler>=0.60.0
line-profiler>=3.5.0
py-spy>=0.3.0
psutil>=5.9.0

# Security and Cryptography
cryptography>=37.0.0
hashlib
secrets

# File and Data Formats
openpyxl>=3.0.0
xlrd>=2.0.0
pyarrow>=8.0.0
parquet>=1.0.0
h5py>=3.7.0
netcdf4>=1.6.0

# Time and Date Handling
pytz>=2022.1
python-dateutil>=2.8.0
arrow>=1.2.0
pendulum>=2.1.0

# Signal Processing
scipy.signal
scipy.fft
scipy.interpolate

# Statistics and Mathematics
statsmodels>=0.13.0
arch>=5.2.0     # Financial econometrics
pykalman>=0.9.0
filterpy>=1.4.0

# Optimization
cvxpy>=1.2.0
pulp>=2.6.0
ortools>=9.3.0

# Natural Language Processing (for sentiment analysis)
nltk>=3.7
textblob>=0.15.3
vaderSentiment>=3.3.2
spacy>=3.4.0

# Image Processing (for chart analysis)
opencv-python>=4.6.0
pillow>=9.0.0
imageio>=2.19.0

# Network and Communication
zeromq>=0.0.0
pika>=1.2.0    # RabbitMQ
kafka-python>=2.0.0
confluent-kafka>=1.8.0

# Cloud and Deployment
boto3>=1.24.0  # AWS
google-cloud-storage>=2.5.0
azure-storage-blob>=12.14.0
docker>=6.0.0
kubernetes>=26.0.0

# System and OS
psutil>=5.9.0
shutil
subprocess
multiprocessing
threading

# Development Tools
jupyter>=1.0.0
ipython>=8.0.0
notebook>=6.4.0
voila>=0.3.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.17.0

# Version Control and CI/CD
gitpython>=3.1.0
semver>=2.13.0

# Optional: GPU Acceleration
# cuda-python>=11.7.0
# cupy>=10.0.0

# Optional: Advanced Analytics
# prophet>=1.1.0  # Facebook time series forecasting
# pmdarima>=1.8.0  # Auto ARIMA
# pycaret>=3.0.0  # Auto ML

# Optional: Blockchain and Crypto
# web3>=5.31.0
# eth-account>=0.6.0
# bitcoin>=1.1.42

# Optional: Alternative Data
# newsapi-python>=0.2.6
# tweepy>=4.10.0
# praw>=7.6.0

# Development and Debugging
ipdb>=0.13.0
debugpy>=1.6.0
rich>=12.0.0
tqdm>=4.64.0

# Note: Some packages may have version conflicts
# Install with: pip install -r requirements.txt
# For development: pip install -r requirements.txt[dev]